LayoutTests/imported/w3c/web-platform-tests/cookies/partitioned-cookies/partitioned-cookies.tentative.https-expected.txt


FAIL Partitioned cookies accessible on the top-level site they are created in via HTTP assert_equals: Expected __Host-pccookistore to be available on the top-level site it was created in expected true but got false
FAIL Partitioned cookies accessible on the top-level site they are created in via DOM assert_equals: Expected __Host-pccookistore to be available on the top-level site it was created in expected true but got false
FAIL Partitioned cookies accessible on the top-level site they are created in via CookieStore promise_test: Unhandled rejection with value: object "NotSupportedError: The operation is not supported."
PASS Cross-site window opened correctly
FAIL Partitioned cookies are not accessible on a different top-level site via HTTP assert_equals: Expected __Host-pchttp to not be available on a different top-level site expected false but got true
FAIL Partitioned cookies are not accessible on a different top-level site via DOM assert_equals: Expected __Host-pchttp to not be available on a different top-level site expected false but got true
FAIL Partitioned cookies are not accessible on a different top-level site via CookieStore promise_test: Unhandled rejection with value: object "NotSupportedError: The operation is not supported."
